--- Assets/Scenes/HealthParent.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: dd94d95fbc8294e4b95ffc90fdf6ab74
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
