+++ Assets/Scenes/MainScene.unity
@@ -1312,43 +1312,43 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 1205676469953892238, guid: a7b77d89355717841be7c25584620f1d, type: 3}
       propertyPath: iron
-      value: 1000
+      value: 500
       objectReference: {fileID: 0}
     - target: {fileID: 1205676469953892238, guid: a7b77d89355717841be7c25584620f1d, type: 3}
       propertyPath: bones
-      value: 1000
+      value: 500
       objectReference: {fileID: 0}
     - target: {fileID: 1205676469953892238, guid: a7b77d89355717841be7c25584620f1d, type: 3}
       propertyPath: diamond
-      value: 1000
+      value: 500
       objectReference: {fileID: 0}
     - target: {fileID: 1205676469953892238, guid: a7b77d89355717841be7c25584620f1d, type: 3}
       propertyPath: demonEgg
-      value: 1000
+      value: 500
       objectReference: {fileID: 0}
     - target: {fileID: 1205676469953892238, guid: a7b77d89355717841be7c25584620f1d, type: 3}
       propertyPath: lavaOrbs
-      value: 1000
+      value: 500
       objectReference: {fileID: 0}
     - target: {fileID: 1205676469953892238, guid: a7b77d89355717841be7c25584620f1d, type: 3}
       propertyPath: obsidian
-      value: 1000
+      value: 500
       objectReference: {fileID: 0}
     - target: {fileID: 1205676469953892238, guid: a7b77d89355717841be7c25584620f1d, type: 3}
       propertyPath: demonCore
-      value: 1000
+      value: 500
       objectReference: {fileID: 0}
     - target: {fileID: 1205676469953892238, guid: a7b77d89355717841be7c25584620f1d, type: 3}
       propertyPath: demonBlood
-      value: 1000
+      value: 500
       objectReference: {fileID: 0}
     - target: {fileID: 1205676469953892238, guid: a7b77d89355717841be7c25584620f1d, type: 3}
       propertyPath: orbsOfAcid
-      value: 1000
+      value: 500
       objectReference: {fileID: 0}
     - target: {fileID: 1205676469953892238, guid: a7b77d89355717841be7c25584620f1d, type: 3}
       propertyPath: rottenFlesh
-      value: 1000
+      value: 500
       objectReference: {fileID: 0}
     - target: {fileID: 2254718921059602131, guid: a7b77d89355717841be7c25584620f1d, type: 3}
       propertyPath: m_Name
