--- Assets/Scripts/CarryOverMusic.cs
@@ -1,17 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class CarryOverMusic : MonoBehaviour
-{
-    private void Awake()
-    {
-        GameObject[] musicObj = GameObject.FindGameObjectsWithTag("GameMusic");
-        if(musicObj.Length > 1)
-        {
-            Destroy(this.gameObject);
-        }
-
-        DontDestroyOnLoad(this.gameObject);
-    }
-}
